Clean export module to minimal APIs for spatialdata writing
